Add files via upload
n conclusion, the CatBoostClassifier was trained and evaluated with a comprehensive set of hyperparameters, achieving a test accuracy of approximately 83.49%. This performance suggests that the model is reasonably effective in predicting the target variable based on the provided features. The training process involved 464 iterations with early stopping applied to prevent overfitting, and the chosen evaluation metric was accuracy.